Move drag to edge canvas to separate file
