+++ applications/conceptual-model-editor/src/app/diagram/reactflow/components/drag-edge-canvas-menu.tsx
@@ -16,8 +16,6 @@ export const useCanvasMenuOptions = () => {
             openCreateClassDialogHandler: () => void;
             openSelectClassesDialogHandler: () => void;
         }) => {
-
-
         const handlersWithTexts: Array<HandlerWithLabel> = [
             { handler: props.openCreateClassDialogHandler, label: "➕ Create new class" },
             { handler: "Horizontal Separator", label: "" },
